Add effect for Arry / Esgred sacrifice
The "sacrifice [the current card] when you control [different card]"
ability of Arry and Esgred is now implemented as an effect. This allows
better consistency with the rest of the engine as well as automatically
taking care of blanking (i.e. marshalling the card during Fortified
Position). The effect takes an optional callback that is fired when the
sacrifice is triggered.